# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AdmitSimple HIPAA-Conscious Admissions CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activities
      parameters:
        - name: inquiryId
          in: query
          schema:
            type: integer
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /activities/{id}:
    put:
      operationId: updateActivity
      tags: [activities]
      summary: Update activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActivityBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Activity:
      type: object
      properties:
        id:
          type: integer
        inquiryId:
          type: integer
          nullable: true
        patientId:
          type: integer
          nullable: true
        userId:
          type: integer
          nullable: true
        userName:
          type: string
          nullable: true
        type:
          type: string
        subject:
          type: string
          nullable: true
        body:
          type: string
          nullable: true
        outcome:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - createdAt
    CreateActivityBody:
      type: object
      properties:
        inquiryId:
          type: integer
        patientId:
          type: integer
        type:
          type: string
        subject:
          type: string
        body:
          type: string
        outcome:
          type: string
        scheduledAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
      required:
        - type
    UpdateActivityBody:
      type: object
      properties:
        type:
          type: string
        subject:
          type: string
        body:
          type: string
        outcome:
          type: string
        scheduledAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...