# lib/api-zod/src/generated/types/createActivityBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AdmitSimple HIPAA-Conscious Admissions CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateActivityBody {
  inquiryId?: number;
  patientId?: number;
  type: string;
  subject?: string;
  body?: string;
  outcome?: string;
  scheduledAt?: Date;
  completedAt?: Date;
}